Changed the distance variable type to float since the location manager is using that by default

+++ app/src/main/java/com/codecamp/bitfit/database/Run.java
@@ -14,14 +14,14 @@ import java.util.UUID;
 
 public class Run extends Workout {
     @Column
-    double distance;
+    float distance;
 
     // Getter and Setter
-    public double getDistanceInKm() {
+    public float getDistanceInKm() {
         return distance;
     }
 
-    public void setDistanceInKm(double distance) {
+    public void setDistanceInKm(float distance) {
         this.distance = distance;
     }
